add comments, make more like autoswap interface, rename "myFirstDapp"
